make standard_xmlns an optional param & c-refactor
The “standard_xmlns” parameter in xml_struct (and __xml_struct__) is now optional.
In case it is not provided, the function __xml_struct__ uses takes now the information
from the first (Knora project-) ontologie it finds within the given resource_set.
This should work most of the time in the KNORA environment, since a project uses
its project ontology and shared ontologies only.
In case it won’t produce the right result in the xml-file, the user is allways allowed to
set the parameter explicitely.